[Virtual Uart] Add Echo Server application
