+++ sw/examples/virtual_uart_echo_server/Makefile
