+++ sw/examples/virtual_uart_echo_server/ld/DEMOSYSTEM_FLASH.ld
