+++ sw/examples/virtual_uart_echo_server/src/startup.s
