+++ sw/host/virtual_uart/src/utils.h
@@ -1,0 +1,12 @@
+// Author: Manuel Maddaluno <manuel.maddaluno@unina.it>
+// Description: Virtual uart utility functions header file
+
+#ifndef UTILS_H__
+#define UTILS_H__
+
+/* Utility functions */
+void disable_buffering();
+void enable_buffering();
+void help(char * ex_name);
+
+#endif
